+++ docs/changelog.rst
@@ -24,7 +24,6 @@ New library methods:
 -  :func:`ocdskit.hierarchy.get_base_classes_via_fca`
 -  :func:`ocdskit.normalize.convert_from_oas3`
 -  :func:`ocdskit.normalize.fix_validation_errors`
--  :func:`ocdskit.normalize.get_definitions_keyword`
 -  :func:`ocdskit.normalize.get_normal_schema`
 -  :func:`ocdskit.normalize.get_schema_hash`
 -  :func:`ocdskit.normalize.hoist_deep_properties`
@@ -32,6 +31,7 @@ New library methods:
 -  :func:`ocdskit.normalize.remove_fields`
 -  :func:`ocdskit.normalize.remove_private_fields`
 -  :func:`ocdskit.normalize.remove_unreachable_definitions`
+-  :func:`ocdskit.util.get_definitions_keyword`
 -  :func:`ocdskit.util.longest_common_subsequence`
 
 Changed